% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{d.alpha}
\alias{d.alpha}
\title{compute derivative of alpha w.r.t. zeta}
\usage{
d.alpha(zeta)
}
\arguments{
\item{zeta:}{a numeric vector of length J-1}
}
\value{
a derivative matrix of size (J-1) x (J-1)
}
\description{
This function computes the derivative of alpha w.r.t. zeta
}
\keyword{internal}
