% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{d.inv.logit}
\alias{d.inv.logit}
\title{Inverse Logit Derivative}
\usage{
d.inv.logit(x)
}
\arguments{
\item{x:}{a numeric vector}
}
\value{
the inverse logit derivative
}
\description{
This function computes derivative of inverse logit at all elements in a numeric vector
}
