% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get.alpha}
\alias{get.alpha}
\title{compute alpha from zeta}
\usage{
get.alpha(zeta)
}
\arguments{
\item{zeta:}{a numeric vector}
}
\value{
alpha vector
}
\description{
This function computes the parameter vector alpha from the vector zeta
}
