% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zeta.R
\name{get.zeta}
\alias{get.zeta}
\title{compute zeta from alpha}
\usage{
get.zeta(alpha)
}
\arguments{
\item{zeta:}{a numeric vector}
}
\value{
alpha vector
}
\description{
This function computes the parameter vector zeta from the vector alpha
}
