% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{inv.logit.d}
\alias{inv.logit.d}
\title{Inverse Logit Derivative}
\usage{
inv.logit.d(x)
}
\arguments{
\item{x:}{a numeric vector}
}
\value{
the inverse logit derivative
}
\description{
This function computes derivative of inverse logit at all elements in a numeric vector
}
