% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv.logit}
\alias{inv.logit}
\title{Inverse logit function}
\usage{
inv.logit(x)
}
\arguments{
\item{x:}{a numeric vector}
}
\value{
the inverse logit values
}
\description{
This function computes the inverse logit of all elements in a numeric vector
}
