% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Ordinal Logistic Regression Log-likelihood Function}
\usage{
loglik(y, x, zeta, beta)
}
\arguments{
\item{y:}{outcome vector; a numeric vector of length n with values in 1, ..., J corresponding to the J ordinal outcome categories}

\item{x:}{covariate matrix; n x p matrix with numeric elements}

\item{zeta:}{used to generate category specific intercepts alpha = sum(exp(zeta)); numeric vector of length J-1}

\item{beta:}{slope parameters; numeric vector of length p}
}
\value{
the log-likelihood value
}
\description{
This function computes the log-likelihood value from the ordinal regression model (with a logit link) given data and parameters
}
