% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{ordreg.alphatozeta}
\alias{ordreg.alphatozeta}
\title{compute zeta from alpha}
\usage{
ordreg.alphatozeta(alpha)
}
\arguments{
\item{zeta:}{a numeric vector}
}
\value{
alpha vector
}
\description{
This function computes the parameter vector zeta from the vector alpha
}
\keyword{internal}
