% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{ordreg.lasso}
\alias{ordreg.lasso}
\title{Ordinal Logistic Regression with LASSO Penalty}
\usage{
ordreg.lasso(formula, data, lambda = 0)
}
\arguments{
\item{formula:}{a symbolic description of the model to be fitted; an object of class \code{"formula"}}

\item{data:}{a data frame containing the variables in the model}

\item{lambda:}{LASSO penalty parameter; a non-negative number; default is \code{lambda = 0}, which corresponds to no penalty term}
}
\value{
a list with the following elements:
\itemize{
\item{alpha: the estimated category specific intercepts for the lowest J-1 outcome categories; a numeric vector of length J-1}
\item{beta: the estimated slopes; a numeric vector of length p}
\item{loglik: the log-likelihood value at convergence (not including the LASSO penalty)}
\item{lasso.penalty: the penalty term evaluated at the estimates for alpha and beta}
}
}
\description{
This function computes the log-likelihood value from the ordinal regression model (with a logit link) given data and parameters
}
