% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{ordreg.predict}
\alias{ordreg.predict}
\title{Ordinal Logistic Regression Prediction}
\usage{
ordreg.predict(alpha, beta, x)
}
\arguments{
\item{alpha:}{category specific intercepts; numeric vector of length J-1}

\item{beta:}{slope parameters; numeric vector of length p}

\item{x:}{covariate matrix; n x p matrix with numeric elements}
}
\value{
a numeric vector of length n with predicted y values in {1, ..., J}
}
\description{
This function computes predicted values given ordinal regression coefficients and a data frame with covariates.
}
