% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{ordreg.zetatoalpha}
\alias{ordreg.zetatoalpha}
\title{compute alpha from zeta}
\usage{
ordreg.zetatoalpha(zeta)
}
\arguments{
\item{zeta:}{a numeric vector}
}
\value{
alpha vector
}
\description{
This function computes the parameter vector alpha from the vector zeta
}
\keyword{internal}
