% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{prox.proj}
\alias{prox.proj}
\title{Proximal Projection Operator}
\usage{
prox.proj(z, lambda, m, ind = NULL)
}
\arguments{
\item{z:}{a numeric vector to be projected}

\item{lambda:}{LASSO penalty parameter; a non-negative number}

\item{m:}{step size; a non-negative number}

\item{ind:}{an optional vector with indices indicating which components of z should be projected; defaults to NULL, in which case all components of z are projected}
}
\value{
a numeric vector of the same length as z
}
\description{
This function computes the proximal projection operator for the LASSO penalty used in the proximal gradient descent algorithm
}
\keyword{internal}
