% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforest_features.R
\name{randomforest.features}
\alias{randomforest.features}
\title{Feature Selection by Ordinal Random Forest Importance}
\usage{
randomforest.features(
  y_name,
  data,
  pValue = 0.01,
  maxRuns = 100,
  doTrace = 0,
  holdHistory = FALSE,
  ...
)
}
\arguments{
\item{y_name:}{character, the factor column in data containing the ordinal y values}

\item{data:}{a dataframe containing the outcome variable and the features to be selected}
}
\value{
A list containing: \describe{
\item{y.unique}{A vector of the trained scores from the ordinal random forest model}
\item{forest}{The fitted ordinal random forest model}
\item{boruta}{The result of the Boruta feature selection}
\item{important.features}{A vector of the names of the important features}
\item{times}{A list of the computation times for the two major steps}
}
}
\description{
Uses the OrdinalForest and Boruta packages to select important features from a dataframe
}
\examples{

dat <- simulate.data(
  n = 200,
  alpha = seq(0.5, 4, length=3),
  beta = c(rep(1,5), rep(0,10)))

result <- randomforest.features("y", dat)
result$important.features

}
