% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim.data}
\alias{sim.data}
\title{Simulate Data from Ordinal Logistic Regression Model}
\usage{
sim.data(formula, n, alpha, beta)
}
\arguments{
\item{formula:}{a symbolic description of the model from which data are generated; an object of class \code{"formula"}}

\item{n:}{number of observations to be simulated: a positive integer}

\item{alpha:}{category specific intercepts for the lowest J-1 outcome categories; a numeric vector of length J-1}

\item{beta:}{slope coefficients; a numeric vector of length p}
}
\value{
a data frame with n rows and the following p + 1 columns
}
\description{
This function simulates a data set from an ordinal regression model
}
