% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{slow.fit}
\alias{slow.fit}
\title{Slow Optimization Procedure}
\usage{
slow.fit(y, x, zeta0, beta0, lambda)
}
\arguments{
\item{y:}{outcome vector; a numeric vector of length n with values in 1, ..., J corresponding to the J ordinal outcome categories
@param x: covariate matrix; n x p matrix with numeric elements
@param zeta0: starting value for zeta; numeric vector of length J-1
@param beta0: starting value for beta; numeric vector of length p
@param lambda: LASSO penalty parameter; a non-negative number

#' @return a list with the following elements:
\itemize{
\itme{zeta: the estimated zeta parameters; a numeric vector of length J-1}
\item{beta: the estimated slopes; a numeric vector of length p}
\item{loglik.val: the log-likelihood value at convergence}
\item{obj.val: the objective function value at convergence}
\item{n.iterations: the number of outer loop iterations in the PGD algorithm}
}}
}
\description{
Slow procedure to fit LASSO-penalized ordinal regression model; only included for purposes of comparing to faster approaches (like PGD).
}
