% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg_lasso.R
\name{ordreg.lasso}
\alias{ordreg.lasso}
\title{Ordinal Logistic Regression with LASSO Penalty}
\usage{
ordreg.lasso(formula, data, lambdas = 0, return.cov = FALSE)
}
\arguments{
\item{formula:}{a symbolic description of the model to be fitted; an object of class \code{"formula"}}

\item{data:}{a data frame containing the variables in the model}

\item{lambdas:}{vector of LASSO penalty parameters; a vector of non-negative numbers; default is 0, which corresponds to no penalization}

\item{return.cov:}{an indicator for whether asymptotic covariance matrix of parameter estimates should be output; defaults to FALSE}
}
\value{
a list with the following elements:
\itemize{
\item{lambdas: vector of LASSO penalty parameters (sorted in descending order)}
\item{alpha: matrix of alpha estimates with each row corresponding to a lambda value}
\item{beta: matrix of beta estimates with each row corresponding to a lambda value}
\itme{n.nonzero: number of nonzero parameters for each lambda value; a vector of non-negative integers}
\item{loglik.val: log-likelihood values at convergence (not including the LASSO penalty) for each lambda value; a numeric vector}
\item{bic: Bayesian information critetion value at convergence for each lambda value; a numeric vector}
\item{predictions: matrix of predicted y values with each row corresponding to a lambda value}
\item{kappa: vector of weighted kappa values for each lambda value}
\item{cov: asymptotic covariance matrix of parameter estimates from unpenalized fit (lambda = 0)}
}
}
\description{
This function fits LASSO-penalized ordinal regression models over a grid of penalty parameters.
}
